+++ source/types/ITheme.ts
@@ -1,0 +1,12 @@
+import { Language } from "../utilities/LanguageService";
+
+/**
+ * Interface for app themes that can be saved
+ */
+export type ITheme = {
+  name: Record<Language, string>;
+  prefix: string;
+  primaryColor: string;
+  secondaryColor: string;
+  cursor: string;
+};
