+++ source/utilities/CacheService.ts
@@ -1,24 +1,21 @@
+import { ITheme } from "../types/ITheme.js";
+import { IThemeOverride } from "../types/IThemeOverides.js";
 import { Dungeon, initiateDungeonMapWithHallways } from "./DungeonService.js";
 import { getTerm, Language } from "./LanguageService.js";
 import { log } from "./LogService.js";
 import { getSettingsData, saveSettingsData } from "./SettingsService.js";
-import { ITheme, IThemeOverride, standardTheme } from "./ThemingService.js";
+import { standardTheme } from "./ThemingService.js";
 
 let cachedDungeon: Dungeon = initiateDungeonMapWithHallways();
-let cachedLanguage: Language = "de";
-let cachedTheme: ITheme = standardTheme;
-let cachedPassword: string =
-  "a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3";
+let cachedLanguage: Language;
+let cachedTheme: ITheme;
+let cachedPassword: string;
 
 loadData();
 
 // #region Getters
 export function getDungeon() {
   return cachedDungeon;
-}
-
-export function renewDungeon() {
-  cachedDungeon = initiateDungeonMapWithHallways();
 }
 
 export function getPassword() {
@@ -35,6 +32,10 @@ export function getTheme() {
 // #endregion
 
 // #region Setters
+export function resetDungeon() {
+  cachedDungeon = initiateDungeonMapWithHallways();
+}
+
 export function setDungeon(dungeon: Dungeon) {
   cachedDungeon = dungeon;
   saveData();
