+++ source/utilities/DungeonService.ts
@@ -128,7 +128,6 @@ export function getDungeonMapVisual() {
     let rowVisual = [[""], [""], [""], [""], [""]];
     for (let col = 0; col < dungeon.size; col++) {
       const miniRoom = getMiniRoomVisual(dungeon.rooms[row][col], row, col);
-      const miniRoom = getMiniRoomVisual(dungeon.rooms[row][col], row, col);
       rowVisual[0].push(miniRoom.split("\n")[0]);
       rowVisual[1].push(miniRoom.split("\n")[1]);
       rowVisual[2].push(miniRoom.split("\n")[2]);
